Add Typescript for Homepage
Add Typescript types for the Home page.
